# Refactor file content searching