- [X] This comment contains a description of changes (with reason)

This report does the following:

- Create a new SearchFile class that contains a line_iterator function.
  - This iterator function caches the filehandle used to read the file.
  - SearchFile is a context manager, and closes the file on exit. 

The previous code would reuse the content lines as much as possible, but would read the file again when not enough content lines were read already. By using a object, the existing filehandle can be kept alive. This prevents reopening and rereading. 

Also by moving all the line reading code into a separate function, the resulting code in `search_file` has become more readable in my opinion. 

The speed uplift is very significant. Tested with ` multiqc --force --no-data-dir --profile-runtime ../test-data/` (The multiqc test-data repo)

Before:
```
|           multiqc | Run took 24.66 seconds
|           multiqc |  - 9.70s: Searching files
|           multiqc |  - 8.40s: Running modules
|           multiqc |  - 5.07s: Compressing report data
```
After:
```
|           multiqc | Run took 20.25 seconds
|           multiqc |  - 5.12s: Searching files
|           multiqc |  - 8.44s: Running modules
|           multiqc |  - 5.11s: Compressing report data
```

A more than 40% reduction in searching files is very welcome. This was benchmarked on a nvme SSD, but I regularly work in a Network Filesystem which is orders of magnitude slower. These speed increases become very significant in such environments.