@vladsavelyev 2024-04-27
This is a really amazing refactoring, thanks so much for this! Love the use of the lazy `line_iterator` method, and encapsulation into the `SearchFile` class. And I always wished to see the search file as a class to enable static type checking and autocompletion. 

Should probably go further and try to put `SearchFile` into `report.files` and expose to `find_log_files()` in all modules. Add type hints, replace the dict keying with object attribute access, and use the `path` property and the line iterator whenever it makes sense. Happy to give it a try in another PR.