@rhpvorderman 2024-04-29
> Should probably go further and try to put SearchFile into report.files and expose to find_log_files() in all modules. Add type hints, replace the dict keying with object attribute access, and use the path property and the line iterator whenever it makes sense. Happy to give it a try in another PR.

Before you do that, I have another PR in the waiting that depends on this PR and makes the file searching another 20% faster. I will open it as soon as this is merged.